Add g:luasyn_nosymboloperator & luaSymbolOperator.
The option g:luasyn_nosymboloperator = 1 can be used to completely disable
highlighting of the Lua symbol operators (that is, all operators except
for the keyword operators "and" and "or").

If that option is not set (the default) then the symbol operators will
use their new own highlighting group "luaSymbolOperator" that is linked
to "luaOperator".