syntax: Add option to display ~= as Not Equal To
When switching back and forth between Lua and other languages, I
sometimes get confused by Lua's unusual use of the ~= instead of !=.

When using ligature fonts like Fira Code, ~= is displayed with Unicode
Asymptotically Equal To which is *very* wrong.

This change resolves both of these issues by making the ~= operator
display as the mathematically familiar Unicode Not Equal To: ≠

It's disabled by default. Enable with:

  let g:lua_syntax_fancynotequal = 1

It cannot be enabled when using lua_syntax_nosymboloperator.

Tested with Penlight's dir and it everything looked right with and
without lua_syntax_fancynotequal.
https://raw.githubusercontent.com/Tieske/Penlight/master/lua/pl/dir.lua